requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
openpyxl
python-dotenv
google-generativeai
